backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMySQL==1.1.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
